istream: Request the minimal HEADER_SIZE and not MAGIC_SIZE
When looking at the header in an istream source, request from the parent
the HEADER_SIZE which contains the magic value and the version.

Else, it is possible to only have the MAGIC_SIZE and thus failing to
read the version.

This has been observed on Riseup email infrastructure leading to a
Permission denied to open the mbox because of the unknown version:

Error: [trees] Unknown version 0. Supporting 1 to 1

Signed-off-by: David Goulet <dgoulet@riseup.net>